+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 
+## 2025-05-30 11:53:51 - Commit d68c07a
+
+- Auto-deploy: 2025-05-30 11:53:48
+
+### Fichiers modifiés:
+- CHANGELOG.md
+- deploy/deploy-status.json
+- scripts/launch.sh
+
+
 ## 2025-05-30 11:52:16 - Commit e3a4db4
 
 - Auto-deploy: 2025-05-30 11:52:10